+++ compiler-plugin/src/main/kotlin/tech/mappie/MappieCommandLineProcessor.kt
@@ -60,6 +60,12 @@ class MappieCommandLineProcessor : CommandLineProcessor {
             valueDescription = "string",
             description = "report directory",
             required = false,
+        ),
+        CliOption(
+            optionName = OPTION_OUTPUT_DIR,
+            valueDescription = "string",
+            description = "output directory",
+            required = true,
         )
     )
 
@@ -73,6 +79,7 @@ class MappieCommandLineProcessor : CommandLineProcessor {
             OPTION_NAMING_CONVENTION -> configuration.put(ARGUMENT_NAMING_CONVENTION, value)
             OPTION_REPORT_ENABLED -> configuration.put(ARGUMENT_REPORT_ENABLED, value.toBooleanStrict())
             OPTION_REPORT_DIR -> configuration.put(ARGUMENT_REPORT_DIR, value)
+            OPTION_OUTPUT_DIR -> configuration.put(ARGUMENT_OUTPUT_DIR, value)
             else -> throw IllegalArgumentException("Unknown option ${option.optionName}")
         }
     }
@@ -86,6 +93,7 @@ class MappieCommandLineProcessor : CommandLineProcessor {
         const val OPTION_NAMING_CONVENTION = "naming-convention"
         const val OPTION_REPORT_ENABLED = "report-enabled"
         const val OPTION_REPORT_DIR = "report-dir"
+        const val OPTION_OUTPUT_DIR = "output-dir"
 
         val ARGUMENT_WARNINGS_AS_ERRORS = CompilerConfigurationKey<Boolean>(OPTION_WARNINGS_AS_ERRORS)
         val ARGUMENT_USE_DEFAULT_ARGUMENTS = CompilerConfigurationKey<Boolean>(OPTION_USE_DEFAULT_ARGUMENTS)
@@ -95,5 +103,6 @@ class MappieCommandLineProcessor : CommandLineProcessor {
         val ARGUMENT_NAMING_CONVENTION = CompilerConfigurationKey<String>(OPTION_NAMING_CONVENTION)
         val ARGUMENT_REPORT_ENABLED = CompilerConfigurationKey<Boolean>(OPTION_REPORT_ENABLED)
         val ARGUMENT_REPORT_DIR = CompilerConfigurationKey<String>(OPTION_REPORT_DIR)
+        val ARGUMENT_OUTPUT_DIR = CompilerConfigurationKey<String>(OPTION_OUTPUT_DIR)
     }
 }