+++ website/src/changelog.md
@@ -2,6 +2,10 @@
 title: "Changelog"
 layout: "layouts/changelog.html"
 changelog:
+  - date: "tbd"
+    title: "v2.3.10-2.4.0"
+    items:
+      - "[#240](https://github.com/Mr-Mappie/mappie/issues/240) Fixed issue with incremental compilation where mappers cannot be found."
   - date: "2026-02-07"
     title: "v2.3.10-2.3.1"
     items: